refactor: e2e, split util.go
Split util.go along resource and function lines
into a bunch of smaller files.
