+++ test/e2e/common/util_node.go
@@ -1,0 +1,73 @@
+package common
+
+// Utility functions for manipulation of nodes.
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os/exec"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
+)
+
+type NodeLocation struct {
+	NodeName     string
+	IPAddress    string
+	MayastorNode bool
+}
+
+// returns vector of populated NodeLocation structs
+func GetNodeLocs() ([]NodeLocation, error) {
+	nodeList := corev1.NodeList{}
+
+	if gTestEnv.K8sClient.List(context.TODO(), &nodeList, &client.ListOptions{}) != nil {
+		return nil, errors.New("failed to list nodes")
+	}
+	NodeLocs := make([]NodeLocation, 0, len(nodeList.Items))
+	for _, k8snode := range nodeList.Items {
+		addrstr := ""
+		namestr := ""
+		mayastorNode := false
+		for label, value := range k8snode.Labels {
+			if label == "openebs.io/engine" && value == "mayastor" {
+				mayastorNode = true
+			}
+		}
+		for _, addr := range k8snode.Status.Addresses {
+			if addr.Type == corev1.NodeInternalIP {
+				addrstr = addr.Address
+			}
+			if addr.Type == corev1.NodeHostName {
+				namestr = addr.Address
+			}
+		}
+		if namestr != "" && addrstr != "" {
+			NodeLocs = append(NodeLocs, NodeLocation{NodeName: namestr, IPAddress: addrstr, MayastorNode: mayastorNode})
+		} else {
+			return nil, errors.New("node lacks expected fields")
+		}
+	}
+	return NodeLocs, nil
+}
+
+// TODO remove dependency on kubectl
+// label is a string in the form "key=value"
+// function still succeeds if label already present
+func LabelNode(nodename string, label string, value string) {
+	labelAssign := fmt.Sprintf("%s=%s", label, value)
+	cmd := exec.Command("kubectl", "label", "node", nodename, labelAssign, "--overwrite=true")
+	cmd.Dir = ""
+	_, err := cmd.CombinedOutput()
+	Expect(err).ToNot(HaveOccurred())
+}
+
+// TODO remove dependency on kubectl
+// function still succeeds if label not present
+func UnlabelNode(nodename string, label string) {
+	cmd := exec.Command("kubectl", "label", "node", nodename, label+"-")
+	cmd.Dir = ""
+	_, err := cmd.CombinedOutput()
+	Expect(err).ToNot(HaveOccurred())
+}
